Fst2/requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
openai
langchain
pinecone-client
